Move PrairieTest auth to ee directory
